+++ core/src/main/resources/META-INF/services/io.questdb.griffin.FunctionFactory
@@ -926,8 +926,8 @@ io.questdb.griffin.engine.functions.str.RPadVarcharVarcharFunctionFactory
 
 # Trim
 io.questdb.griffin.engine.functions.str.TrimStrFunctionFactory
-io.questdb.griffin.engine.functions.str.LTrimFunctionFactory
-io.questdb.griffin.engine.functions.str.RTrimFunctionFactory
+io.questdb.griffin.engine.functions.str.LTrimStrFunctionFactory
+io.questdb.griffin.engine.functions.str.RTrimStrFunctionFactory
 io.questdb.griffin.engine.functions.str.TrimVarcharFunctionFactory
 io.questdb.griffin.engine.functions.str.LTrimVarcharFunctionFactory
 io.questdb.griffin.engine.functions.str.RTrimVarcharFunctionFactory
